TSTS: Consolidate groupby any, all
